Work so that now the main page and the contribute page work. currently sharing is unsupported
