--- nextjs/components/index.css
@@ -1,123 +1,0 @@
-body {
-    font: 14px "Century Gothic", Futura, sans-serif;
-    margin: 20px;
-}
-
-#main, #share-link-main {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-input {
-    justify-content: center;
-    margin: 20px;
-    padding: 15px;
-    border-radius: 10px;
-    width: 50vw;
-}
-
-form {
-    display: flex;
-    flex-direction: column;
-}
-
-#time {
-    display: flex;
-    flex-direction: row;
-    width: 80vw;
-}
-
-#days {
-    display: flex;
-    flex-direction: column;
-}
-
-#days input {
-    width: 20px;
-}
-
-.class_ input {
-    width: 20px;
-}
-
-#canvas {
-    display: block;
-}
-
-body {
-    margin: 0;
-}
-
-/* Header */
-
-:root {
-    --header-color: #1F2016;
-    --header-text: #E2E3C5;
-}
-
-header {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    padding: 15px;
-    margin: 0px 0px 15px 0px;
-    background-color: var(--header-color);
-    color: var(--header-text);
-}
-
-header a {
-    text-decoration: none;
-    color: var(--header-text);
-}
-
-/* Term tri-select feature css */
-
-:root {
-    --button-down: grey;
-    --button: rgb(180, 180, 180);
-    --button-border: rgb(214, 214, 214);
-}
-
-#term-selector {
-    display: flex;
-    flex-direction: row;
-}
-
-.term {
-    border-radius: 10px;
-    padding: 10px;
-    margin: 10px 10px 10px 10px;
-    background-color: var(--button);
-}
-
-.term:active {
-    background-color: var(--button-down);
-}
-
-.mark-selected {
-    background-color: var(--button-down);
-    border: 5px solid var(--button-border);
-}
-
-.mark-deselected {
-    border: 5px solid var(--button-border);
-    background-color: var(--button);
-}
-
-/* Contribute page */
-
-#contribute {
-    padding: 20px;
-}
-
-#contribute a {
-    text-decoration: none;
-    text-decoration-color: blue;
-}
-
-#contribute a:hover {
-    text-decoration:underline;
-    text-decoration-color: red;
-}
-
